requirements.txt
absl-py==0.11.0
appdirs==1.4.3
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
braceexpand==0.1.6
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cmake==3.18.4.post1
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gym==0.10.9
h5py==2.10.0
habitat-sim==0.1.7
html5lib==1.0.1
idna==2.10
ifcfg==0.21
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
ipaddr==2.2.0
joblib==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
lmdb==1.0.0
lockfile==0.12.2
magnum==0.0.0
Markdown==3.3.3
matplotlib==3.3.4
mock==4.0.3
moviepy==1.0.3
msgpack==0.6.2
networkx==2.5
numba==0.52.0
numpy==1.19.5
numpy-quaternion==2020.11.2.17.0.49
oauthlib==3.1.0
objectio==0.2.29
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
plyfile==0.7.2
proglog==0.1.9
progress==1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytoml==0.1.21
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scikit-fmm==0.0.7
scikit-image==0.17.2
scikit-learn==0.23.1
scipy
simplejson==3.17.2
six==1.15.0
smmap==3.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile
torch==1.7.1
torchgeometry==0.1.2
torchvision==0.8.2
tqdm==4.56.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
webdataset==0.1.40
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
zipp==3.4.0
